Backend/requirements.txt
flask
joblib
numpy
pandas
requests
scikit-learn
flask-cors
python-dotenv